feat: implement tournament evaluation module with jury assignment, submission scoring, and frontend integration
