+++ backend/backend/urls.py
@@ -7,4 +7,5 @@ urlpatterns = [
     path('api/accounts/', include('accounts.urls')),
     path('api/teams/', include('teams.urls')),
     path('api/tournaments/', include('tournaments.urls')),
+    path("api/evaluation/", include("evaluation.urls")),
 ]
