+++ backend/evaluation/urls.py
@@ -1,0 +1,14 @@
+from django.urls import path
+from .views import (
+    JuryAssignmentListView,
+    JuryEvaluationCreateView,
+    JuryEvaluationDetailView,
+    AdminRoundAssignmentView,
+)
+
+urlpatterns = [
+    path('assignments/', JuryAssignmentListView.as_view(), name='jury_assignments'),
+    path('evaluate/<int:assignment_id>/', JuryEvaluationDetailView.as_view(), name='jury_evaluate'),
+    path('evaluate/', JuryEvaluationCreateView.as_view(), name='jury_evaluate_create'),
+    path('rounds/<int:pk>/assign-jury/', AdminRoundAssignmentView.as_view(), name='round_assign_jury'),
+]
