+++ backend/tournaments/urls.py
@@ -1,10 +1,7 @@
 from django.urls import path
 
 from .views import (
-    AdminRoundAssignmentView,
     CurrentTaskView,
-    JuryAssignmentListView,
-    JuryEvaluationCreateView, JuryEvaluationDetailView,
     RoundCloseSubmissionsView,
     RoundDetailView,
     RoundListCreateView,
@@ -36,11 +33,7 @@ urlpatterns = [
     path('rounds/<int:pk>/start/', RoundStartView.as_view(), name='round_start'),
     path('rounds/<int:pk>/close-submissions/', RoundCloseSubmissionsView.as_view(), name='round_close_submissions'),
     path('rounds/<int:pk>/mark-evaluated/', RoundMarkEvaluatedView.as_view(), name='round_mark_evaluated'),
-    path('rounds/<int:pk>/assign-jury/', AdminRoundAssignmentView.as_view(), name='round_assign_jury'),
     path('submissions/', SubmissionListCreateView.as_view(), name='submissions'),
     path('submissions/<int:pk>/', SubmissionDetailView.as_view(), name='submission_detail'),
     path('current-task/', CurrentTaskView.as_view(), name='current_task'),
-    path('jury/assignments/', JuryAssignmentListView.as_view(), name='jury_assignments'),
-    path('jury/evaluate/<int:assignment_id>/', JuryEvaluationDetailView.as_view(), name='jury_evaluate'),
-    path('jury/evaluate/', JuryEvaluationCreateView.as_view(), name='jury_evaluate_create'),
 ]
